Refactored
Made a ../util/hooks folder, moved the navigation utilities there, and changed the imports to match
