+++ src/components/AppBar.tsx
@@ -15,7 +15,7 @@ import { Stack } from '@mui/material';
 import { useAuth } from '../util/auth/useAuth';
 import { useNavigate } from 'react-router-dom';
 import PenguinIcon from '../assets/penguins/penguin-icon.tsx';
-import {useSignUpNavigation, useLoginNavigation} from '../util/navigationUtilities.ts';
+import {useSignUpNavigation, useLoginNavigation} from '../util/hooks/navigationUtilities.ts';
 
 
 const pages = ['Projects', 'Tickets'];
